Add fist draft test pt2
